organizing and improving nip17 and nip59
